requirements.txt
flask==3.0.2
flask-cors==4.0.0
werkzeug==3.0.1
mistralai==1.6.0
google-generativeai==0.3.2
python-dotenv==1.0.1
markdown==3.5.2
pypandoc==1.12 